Moves ICastableHelpers to its own file
