+++ src/System.Private.CoreLib/shared/System/Runtime/CompilerServices/ICastable.cs
@@ -58,24 +58,4 @@ namespace System.Runtime.CompilerServices
         // IsInstanceOfInterface.
         RuntimeTypeHandle GetImplType(RuntimeTypeHandle interfaceType);
     }
-
-#if CORECLR
-    /// <summary>
-    /// Helpers that allows VM to call into ICastable methods without having to deal with RuntimeTypeHandle.
-    /// RuntimeTypeHandle is a struct and is always passed in stack in x86, which our VM call helpers don't
-    /// particularly like.
-    /// </summary>
-    internal class ICastableHelpers
-    {
-        internal static bool IsInstanceOfInterface(ICastable castable, RuntimeType type, out Exception castError)
-        {
-            return castable.IsInstanceOfInterface(new RuntimeTypeHandle(type), out castError);
-        }
-
-        internal static RuntimeType GetImplType(ICastable castable, RuntimeType interfaceType)
-        {
-            return castable.GetImplType(new RuntimeTypeHandle(interfaceType)).GetRuntimeType();
-        }
-    }
-#endif
 }
